Move patch related tests to a separate file
